api/app/requirements.txt
anyio==3.6.1
asgiref==3.5.2
cffi==1.15.0
click==8.1.3
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
greenlet==0.4.13
gunicorn==20.1.0
h11==0.13.0
hpy==0.0.3
idna==3.3
itsdangerous==2.1.2
motor==2.5.1
passlib==1.7.4
pydantic==1.10.1
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.3
readline==6.2.4.1
sniffio==1.3.0
starlette==0.16.0
typing_extensions==4.3.0
ulid-py==1.1.0
uvicorn==0.14.0
